+++ README.md
@@ -1,1 +1,3 @@
 # AxieBot
+- Channel Id is in setLeaderBoard.js
+- Bot token is in startBot.js
